make front50 compatible with asgard, download java 8 and set as default

+++ micronolith/scripts/postInstall.sh
@@ -1,8 +1,10 @@
 # Install Java 8
-curl -L -b "oraclelicense=a" http://download.oracle.com/otn-pub/java/jdk/8u25-b17/jdk-8u25-linux-x64.tar.gz > /tmp/jdk8.tar.gz
-tar -zxvf /tmp/jdk8.tar.gz -C /usr/lib/jvm
-rm /etc/alternatives/java
-ln -s /usr/lib/jvm/jdk1.8.0_25/bin/java /etc/alternatives/java
+if [ ! -d "/usr/lib/jvm/jdk1.8.0_25" ]; then
+  curl -L -b "oraclelicense=a" http://download.oracle.com/otn-pub/java/jdk/8u25-b17/jdk-8u25-linux-x64.tar.gz > /tmp/jdk8.tar.gz
+  tar -zxvf /tmp/jdk8.tar.gz -C /usr/lib/jvm
+  rm /etc/alternatives/java
+  ln -s /usr/lib/jvm/jdk1.8.0_25/bin/java /etc/alternatives/java
+fi
 
 # Fix Front50
 if [ -f "/opt/front50/bin/front50-web" ]; then
@@ -10,8 +12,7 @@ if [ -f "/opt/front50/bin/front50-web" ]; then
 fi
 
 # Move configs to a common location; makes for easier management.
-# NOT FRONT50's, since we're explcitly disabling cassandra, we'll overwrite it.
-for svc in kato mort orca oort gate;
+for svc in kato mort oort orca gate;
 do
   if [ ! -L /opt/$svc/config/$svc.yml ]; then
     mv /opt/$svc/config/$svc.yml /opt/spinnaker/config/
@@ -26,7 +27,9 @@ mv /var/www/scripts/local.js /var/www/scripts/settings.js
 service redis-server restart
 
 # Remove default site
-rm /etc/apache2/sites-enabled/000-default.conf
+if [ -L /etc/apache2/sites-enabled/000-default.conf ]; then
+  rm /etc/apache2/sites-enabled/000-default.conf
+fi
 
 # Configure ProxyPass
 a2enmod proxy_http
